requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
attrs==25.3.0
frozenlist==1.7.0
idna==3.10
mjpeg-streamer==2024.2.8
multidict==6.5.1
netifaces==0.11.0
numpy==2.3.1
opencv-python==4.11.0.86
propcache==0.3.2
pygame==2.6.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.17
PyQt5_sip==12.17.0
pyserial==3.5
yarl==1.20.1
